pyproject.toml
[project]
name = "visionary-lab"
version = "0.1.0"
description = "Create and manage creative content"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-identity>=1.21.0",
    "azure-storage-blob>=12.25.1",
    "fastapi[standard]>=0.115.12",
    "ipykernel>=6.29.5",
    "openai>=1.70.0",
    "opencv-python>=4.11.0.86",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "pydantic-settings>=2.8.1",
    "pytest-asyncio>=0.26.0",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "replicate>=1.0.4",
    "tabulate>=0.9.0",
    "uvicorn>=0.34.0",
]
